refactor: table components
