+++ packages/components/table/src/CompoundTable.tsx
@@ -1,8 +1,8 @@
 import { Table as OriginalTable } from './Table';
-import { TableBody } from './TableBody';
-import { TableCell } from './TableCell';
-import { TableHead } from './TableHead';
-import { TableRow } from './TableRow';
+import { TableBody } from './TableBody/TableBody';
+import { TableCell } from './TableCell/TableCell';
+import { TableHead } from './TableHead/TableHead';
+import { TableRow } from './TableRow/TableRow';
 
 type CompoundTable = typeof OriginalTable & {
   Row: typeof TableRow;
