+++ packages/components/table/src/Table.tsx
@@ -1,4 +1,4 @@
-import { cx, css } from 'emotion';
+import { cx } from 'emotion';
 import React, { forwardRef } from 'react';
 import {
   Box,
@@ -6,31 +6,7 @@ import {
   type PropsWithHTMLElement,
   type ExpandProps,
 } from '@contentful/f36-core';
-import tokens from '@contentful/f36-tokens';
-
-const getStyles = () => {
-  return {
-    inline: css({
-      borderRadius: tokens.borderRadiusMedium,
-      boxShadow: `0 0 0 1px ${tokens.gray200}`,
-      'th:first-child': {
-        borderTopLeftRadius: tokens.borderRadiusMedium,
-      },
-      'th:last-child': {
-        borderTopRightRadius: tokens.borderRadiusMedium,
-      },
-      'tr:last-child td:first-child': {
-        borderBottomLeftRadius: tokens.borderRadiusMedium,
-      },
-      'tr:last-child td:last-child': {
-        borderBottomRightRadius: tokens.borderRadiusMedium,
-      },
-    }),
-    root: css({
-      width: '100%',
-    }),
-  };
-};
+import { getTableStyles } from './Table.styles';
 
 export type TableInternalProps = CommonProps & {
   layout?: 'inline' | 'embedded';
@@ -49,7 +25,7 @@ export const Table = forwardRef<HTMLTableElement, ExpandProps<TableProps>>(
     },
     forwardedRef,
   ) => {
-    const styles = getStyles();
+    const styles = getTableStyles();
     return (
       <Box
         cellPadding="0"
