+++ packages/components/table/src/TableCell/TableCell.tsx
@@ -1,0 +1,70 @@
+import { cx } from 'emotion';
+import React, { forwardRef } from 'react';
+import {
+  Box,
+  type CommonProps,
+  type PropsWithHTMLElement,
+  type ExpandProps,
+} from '@contentful/f36-core';
+
+import { useTableCellContext } from './TableCellContext';
+import { getTableCellStyles } from './TableCell.styles';
+
+export const sortingDirections = {
+  asc: 'asc',
+  desc: 'desc',
+};
+
+export type TableCellSorting = keyof typeof sortingDirections | boolean;
+
+export type TableCellInternalProps = CommonProps & {
+  align?: 'center' | 'left' | 'right';
+  sorting?: TableCellSorting;
+  width?: string | number;
+  children?: React.ReactNode;
+};
+
+export type TableCellProps = PropsWithHTMLElement<
+  TableCellInternalProps,
+  'th' | 'td'
+>;
+
+export const TableCell = forwardRef<
+  HTMLTableCellElement,
+  ExpandProps<TableCellProps>
+>(
+  (
+    {
+      align = 'left',
+      children,
+      className,
+      sorting = false as TableCellSorting,
+      testId = 'cf-ui-table-cell',
+      ...otherProps
+    },
+    forwardedRef,
+  ) => {
+    const { as, name: context, offsetTop } = useTableCellContext();
+
+    const isTableHead = context === 'head';
+    const styles = getTableCellStyles({ isTableHead, sorting, align });
+
+    return (
+      <Box
+        {...otherProps}
+        as={as}
+        className={cx(styles.container, className)}
+        ref={forwardedRef}
+        style={{
+          ...otherProps.style,
+          top: offsetTop || undefined,
+        }}
+        testId={testId}
+      >
+        {children}
+      </Box>
+    );
+  },
+);
+
+TableCell.displayName = 'TableCell';
