+++ packages/components/table/src/TableHead/TableHead.styles.ts
@@ -1,0 +1,12 @@
+import { css } from 'emotion';
+import tokens from '@contentful/f36-tokens';
+
+export const getTableHeadStyles = () => ({
+  sticky: css({
+    th: {
+      position: 'sticky',
+      top: 0,
+      zIndex: tokens.zIndexDefault,
+    },
+  }),
+});
