+++ packages/components/table/src/index.ts
@@ -1,11 +1,11 @@
 export { Table } from './CompoundTable';
 export type { TableProps } from './Table';
-export { TableBody } from './TableBody';
-export type { TableBodyProps } from './TableBody';
-export { TableCell } from './TableCell';
-export type { TableCellProps } from './TableCell';
-export { contextOptions, TableCellContext } from './tableCellContext';
-export { TableHead } from './TableHead';
-export type { TableHeadProps } from './TableHead';
-export { TableRow } from './TableRow';
-export type { TableRowProps } from './TableRow';
+export { TableBody } from './TableBody/TableBody';
+export type { TableBodyProps } from './TableBody/TableBody';
+export { TableCell } from './TableCell/TableCell';
+export type { TableCellProps } from './TableCell/TableCell';
+export { contextOptions, TableCellContext } from './TableCell/TableCellContext';
+export { TableHead } from './TableHead/TableHead';
+export type { TableHeadProps } from './TableHead/TableHead';
+export { TableRow } from './TableRow/TableRow';
+export type { TableRowProps } from './TableRow/TableRow';
